qianfan_autotest/requirements.txt
arrow==0.10.0
futures==3.0.5
beautifulsoup4==4.5.3
PyStaticConfiguration==0.10.3
toolz==0.8.2
pytest==3.2.0
pytest_allure_adaptor==1.7.8
pytest-html==1.16.0
selenium==3.5.0
pytractor==0.2.1
Pillow==4.2.1
pytesseract==0.1.7
requests==2.18.4
jinja2==2.10
pymysql==0.7.9
peewee==2.10.2
tablib==0.11.2
jmespath==0.9.2
lxml==3.8.0
PyYAML==3.12
jsonobject==0.7.1
six==1.10.0
peewee==2.8.5
xlrd==1.1.0
openpyxl==2.4.9
XlsxWriter==1.0.2
pandas==0.21.0
ddt==1.1.1
mock
